backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-auth0==0.0.3
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
idna==3.7
psycopg2-binary==2.9.9
pyasn1==0.6.0
PyJWT==2.9.0
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
